% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRShiny2.R
\name{MLRShiny2}
\alias{MLRShiny2}
\title{Launch MLRShiny2 Interface}
\usage{
MLRShiny2()
}
\value{
Nothing
}
\description{
MLRShiny2() loads interactive user interface built using R shiny.
}
\details{
Start MLRShiny2

The interactive user interface is to provide an easy way for people who are learning Multiple Linear Regression. Includes example data for testing out a few example analysis.
}
\examples{
if(interactive()){
library(shiny)
MLRShiny2()
}
}
\keyword{MLRShiny2}
